Moved some of the paste test code out to a support file, ready to be re-used in some new tests.
